# docs/openapi.yaml
---
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  "/users/me":
    get:
      summary: Get User
      operationId: get_user_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
      security:
      - traO: []
  "/users/groups":
    get:
      summary: Get User Groups
      operationId: get_user_groups_users_groups_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/Group"
                type: array
                title: Response Get User Groups Users Groups Get
      security:
      - traO: []
  "/groups/{group_id}":
    get:
      summary: Get Group
      operationId: get_group_groups__group_id__get
      security:
      - traO: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/GroupDetails"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/groups/{group_id}/tasks":
    get:
      summary: Get Group Tasks
      operationId: get_group_tasks_groups__group_id__tasks_get
      security:
      - traO: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TaskDetails"
                title: Response Get Group Tasks Groups  Group Id  Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/tasks":
    post:
      summary: Create Task
      operationId: create_task_tasks_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTaskReqDTO"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TaskDetails"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - traO: []
  "/tasks/{task_id}":
    delete:
      summary: Delete Task
      operationId: delete_task_tasks__task_id__delete
      security:
      - traO: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/tasks/{task_id}/title":
    patch:
      summary: Patch Task Title
      operationId: patch_task_title_tasks__task_id__title_patch
      security:
      - traO: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      - name: title
        in: query
        required: true
        schema:
          type: string
          title: Title
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/tasks/{task_id}/content":
    patch:
      summary: Patch Task Content
      operationId: patch_task_content_tasks__task_id__content_patch
      security:
      - traO: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      - name: content
        in: query
        required: true
        schema:
          type: string
          title: Content
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/tasks/{task_id}/due_date":
    patch:
      summary: Patch Task Due Date
      operationId: patch_task_due_date_tasks__task_id__due_date_patch
      security:
      - traO: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      - name: due_date
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Due Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/tasks/{task_id}/assignees":
    patch:
      summary: Put Task Assignee
      operationId: put_task_assignee_tasks__task_id__assignees_patch
      security:
      - traO: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
              title: User Ids
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/tasks/{task_id}/labels":
    patch:
      summary: Put Task Assignee
      operationId: put_task_assignee_tasks__task_id__labels_patch
      security:
      - traO: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
              title: Label Ids
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/labels":
    post:
      summary: Create Label
      operationId: create_label_labels_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LabelCreate"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Label"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - traO: []
  "/labels/{label_id}":
    patch:
      summary: Edit Label
      operationId: edit_label_labels__label_id__patch
      security:
      - traO: []
      parameters:
      - name: label_id
        in: path
        required: true
        schema:
          type: string
          title: Label Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LabelUpdate"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Label"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Label
      operationId: delete_label_labels__label_id__delete
      security:
      - traO: []
      parameters:
      - name: label_id
        in: path
        required: true
        schema:
          type: string
          title: Label Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
components:
  schemas:
    CreateTaskReqDTO:
      properties:
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        message_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Message Id
        due_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due Date
        group_id:
          type: string
          title: Group Id
        assigned_user_ids:
          items:
            type: string
          type: array
          title: Assigned User Ids
        label_ids:
          items:
            type: string
          type: array
          title: Label Ids
      type: object
      required:
      - title
      - content
      - group_id
      - assigned_user_ids
      - label_ids
      title: CreateTaskReqDTO
    Group:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        remind_channel_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Remind Channel Id
        periodic_remind_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Periodic Remind At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - remind_channel_id
      - periodic_remind_at
      - created_at
      - updated_at
      title: Group
    GroupDetails:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        remind_channel_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Remind Channel Id
        periodic_remind_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Periodic Remind At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
      type: object
      required:
      - id
      - name
      - remind_channel_id
      - periodic_remind_at
      - created_at
      - updated_at
      - user_ids
      title: GroupDetails
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Label:
      properties:
        name:
          type: string
          title: Name
        id:
          type: string
          title: Id
        group_id:
          type: string
          title: Group Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - name
      - id
      - group_id
      - created_at
      - updated_at
      title: Label
    LabelCreate:
      properties:
        name:
          type: string
          title: Name
        group_id:
          type: string
          title: Group Id
      type: object
      required:
      - name
      - group_id
      title: LabelCreate
    LabelUpdate:
      properties:
        name:
          type: string
          title: Name
      type: object
      required:
      - name
      title: LabelUpdate
    TaskDetails:
      properties:
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        message_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Message Id
        due_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due Date
        id:
          type: string
          title: Id
        group_id:
          type: string
          title: Group Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        labels:
          items:
            "$ref": "#/components/schemas/Label"
          type: array
          title: Labels
        assigned_users:
          items:
            "$ref": "#/components/schemas/User"
          type: array
          title: Assigned Users
      type: object
      required:
      - title
      - content
      - id
      - group_id
      - created_at
      - updated_at
      - labels
      - assigned_users
      title: TaskDetails
    User:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        remind_channel_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Remind Channel Id
        periodic_remind_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Periodic Remind At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - created_at
      - updated_at
      title: User
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    traO:
      type: apiKey
      in: header
      name: X-Forwarded-User